requirements.txt
flask
PyPDF2
random2
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
openai
firebase-admin
google-cloud
